% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated_binomial.R
\name{get_zib_ml}
\alias{get_zib_ml}
\title{Estimate Probability of Occupancy and Conditional Probability of Detection under a Zero-Inflated Binomial model}
\usage{
get_zib_ml(data, ci = 0.95)
}
\arguments{
\item{data}{Data.frame with at least two columns: one named obs_counts with the observed counts at each site, and once called n_tests with the number of tests per sites}

\item{ci}{Value for confidence interval (between 0 and 1)}
}
\value{
A data.frame with maximum likelihood estimates for probability of occupancy and probability of detection given site is occupied, plus confidence interval around the estimates.
}
\description{
Maximum likelihood estimates for probability of occupancy and probability of
detection coditional on site being occupied using the \code{get_zib_likelihood}
function.
}
\examples{
data <- maldetectr::sim_zib_data( n_sites = 10, n_samples = 5, p_oc = 0.2, p_dt = 0.8 )
get_zib_ml( data = data, ci = 0.95 )

}
\seealso{
\code{\link{get_zib_likelihood}}
}

