% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated_binomial.R
\name{sim_zib_data}
\alias{sim_zib_data}
\title{Simulate zero-inflated binomial data}
\usage{
sim_zib_data(n_sites, n_samples, p_oc, p_dt, seed = 42)
}
\arguments{
\item{n_sites}{Number of sites to simulate}

\item{n_samples}{Number of observations at each site}

\item{p_oc}{Probability of occupancy at site}

\item{p_dt}{Probability of detection at site conditional on site being occupied}

\item{seed}{Set the seed for reproduciability}
}
\value{
A data.frame of simulated data with \code{n_sites} rows and three columns. An \code{occupied_sites} column with 0 if site is not occupied and 1 if occupied; an \code{n_tests} column, which indicates the number of times each site was sampled (all values in this column will be equal to \code{n_samples}, and will be used in calculating the likelihood); and an \code{obs_counts} column, which has the number of times a detection occurred (values will be 0 to a maximum of \code{n_smaples})
}
\description{
Simulate zero-inflated binomial data
}
\examples{
maldetectr::sim_zib_data( n_sites = 10, n_samples = 5, p_oc = 0.2, p_dt = 0.8 )
}

